+++ internal/test/parallel_test.go
@@ -6,8 +6,6 @@ import (
 	"sync"
 	"testing"
 
-	"fmt"
-
 	"github.com/declan94/cfcryptfs/corecrypter"
 )
 
@@ -39,7 +37,6 @@ func TestParallelWrite(t *testing.T) {
 			} else {
 				part = text[begin:end]
 			}
-			fmt.Printf("No.%d: len: %d\n", i, len(part))
 			n, err := fd.WriteAt(part, int64(begin))
 			if err != nil {
 				t.Errorf("Write failed: %v", err)
@@ -59,7 +56,7 @@ func TestParallelWrite(t *testing.T) {
 	fd2.ReadAt(text2, 0)
 	if !bytes.Equal(text, text2) {
 		t.Error("Context not matched")
-		t.Errorf("truth: %v", text)
-		t.Errorf("resul: %v", text2)
+		// t.Errorf("truth: %v", text)
+		// t.Errorf("resul: %v", text2)
 	}
 }